requirements.txt
asgiref==3.5.2
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.1.1
django-debug-toolbar==3.2
django-embed-video==1.4.1
idna==3.3
mysqlclient==2.1.1
Pillow==9.0.1
pytz==2020.5
requests==2.27.1
sqlparse==0.2.4
urllib3==1.26.8
